refactor(shared): 归位 JSON 安全截断工具
